Refactor Name Space
Something I have been thinking about doing for a while is just switching
my name spaces to PEP 8, but I never wanted to be inconsistent with
other distributions (also lazy). It may not be completely PEP 8, but I
am trying to make it a little closer. Will get to other distributions
later.